LayoutTests/svg/hittest/text-vertical.html

<!DOCTYPE html>
<title>elementFromPoint(...) on vertical &lt;svg:text></title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/ahem.js"></script>
<style>
body, html {
  padding: 0;
  margin: 0;
}
</style>
<svg width="400" height="400">
  <text x="200" y="50" writing-mode="tb" font-family="Ahem" font-size="50">MMMMMMMM</text>
</svg>
<script>
test(function() {
  var textElement = document.querySelector('text');
  for (var y = 75; y < 400; y += 50)
      assert_equals(document.elementFromPoint(200, y), textElement, "element @ (200," + y + ")");
});
</script>
